[DoNotMerge] ntrdma: measure effect of user space driver
Measure an upper bound on the effect of the user space driver.  This
measures the time from when a recv work request is completed until the
next send work request is posted.  On the server side of the qperf
rc_lat test, this measures the time it takes for the server application
to wake up, poll the completed receive, and respond by posting a send.

The server side can do no better than respond instantaneously, so the
time measured here is an upper bound on the effect that the user space
driver and application might have on the total response time, and
excludes network latency.  Some of the latency might be attributable to
the non-optimized ntrdma user space driver, and could be eliminated by
sharing data structures between user space and kernel space.  This test
measures an upper bound on the latency, but not the precise latency that
is attributable to the non-optimized ntrdma user space driver.

Example:

(server)$ numactl -N 1 -m 1 -- qperf
(client)$ numactl -N 1 -m 1 -- qperf peer rc_lat rc_lat rc_lat
rc_lat:
    latency  =  15.8 us
rc_lat:
    latency  =  15.3 us
rc_lat:
    latency  =  15.4 us

(server)$ dmesg | grep abh
[ 1989.396102] ntrdma 0000:80:03.0: abh time 397171095ns
[ 1989.404184] ntrdma 0000:80:03.0: abh count 63401
[ 1991.838683] ntrdma 0000:80:03.0: abh time 410558586ns
[ 1991.846762] ntrdma 0000:80:03.0: abh count 65169
[ 1994.281167] ntrdma 0000:80:03.0: abh time 412684452ns
[ 1994.289248] ntrdma 0000:80:03.0: abh count 64837

Dividing time by count gives us an average latency per round trip from
the server side recv completion to the next posted send work request.

Qperf measured a latency of about 16us total.  The upper bound latency
that might be attributable to the user space driver is about 6.3us.
This is only latency on the server side, so lets estimate 12us, rounding
down because this is an upper bound, user space driver latency on both
sides.  The proportion of the latency contributed by the user space
driver is 12/16 or 75%.  If we can eliminate the latency attributable to
the user space. Subtracting user space latency, the round trip time
might be improved to a lower bound of 4us.

Again, these are only upper and lower bounds to be used for estimation.
They indicate "we should not expect an effect larger than X" by
optimizing the user space driver.